[#ARR-100] XML/Detail View broken

+++ dcm4chee/dcm4chee-arr/trunk/dcm4chee-arr-seam-ejb/src/main/java/org/dcm4chee/arr/seam/ejb/AuditRecordAccessBean.java
@@ -56,7 +56,7 @@ public class AuditRecordAccessBean implements AuditRecordAccessLocal {
     @PersistenceContext(unitName="dcm4chee-arr")
     private EntityManager enitityManager;
         
-    public AuditRecord findAuditRecord(int pk) {
+    public AuditRecord findAuditRecord(long pk) {
         return enitityManager.find(AuditRecord.class, pk);
     }
 }
