+++ dcm4chee/dcm4chee-arr/trunk/dcm4chee-arr-seam-ejb/src/main/java/org/dcm4chee/arr/seam/ejb/AuditRecordAccessLocal.java
@@ -51,6 +51,6 @@ import org.dcm4chee.arr.entities.AuditRecord;
 @Local
 public interface AuditRecordAccessLocal {
     
-    public AuditRecord findAuditRecord(int pk);
+    public AuditRecord findAuditRecord(long pk);
 
 }
